+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.2.4
 
+* `Versions`
+    * `Serialization`: `1.0.0` -> `1.0.1`
+
 ## 0.2.3
 
 * `Versions`